+++ plugins/bl/regionalinvestment/updates/builder_table_update_bl_regionalinvestment_communities_3.php
@@ -1,4 +1,6 @@
-<?php namespace BL\RegionalInvestment\Updates;
+<?php
+
+namespace BL\RegionalInvestment\Updates;
 
 use Schema;
 use October\Rain\Database\Updates\Migration;
@@ -7,17 +9,19 @@ class BuilderTableUpdateBlRegionalinvestmentCommunities3 extends Migration
 {
     public function up()
     {
-        Schema::table('bl_regionalinvestment_communities', function($table)
-        {
-            $table->dropColumn('map_id');
-        });
+        if (Schema::hasTable('bl_regionalinvestment_communities') && Schema::hasColumn('bl_regionalinvestment_communities', 'map_id')) {
+            Schema::table('bl_regionalinvestment_communities', function ($table) {
+                $table->dropColumn('map_id');
+            });
+        }
     }
-    
+
     public function down()
     {
-        Schema::table('bl_regionalinvestment_communities', function($table)
-        {
-            $table->bigInteger('map_id')->nullable()->unsigned();
-        });
+        if (Schema::hasTable('bl_regionalinvestment_communities') && !Schema::hasColumn('bl_regionalinvestment_communities', 'map_id')) {
+            Schema::table('bl_regionalinvestment_communities', function ($table) {
+                $table->bigInteger('map_id')->nullable()->unsigned();
+            });
+        }
     }
 }
