+++ plugins/bl/regionalinvestment/updates/builder_table_update_bl_regionalinvestment_points_of_interest_2.php
@@ -7,17 +7,21 @@ class BuilderTableUpdateBlRegionalinvestmentPointsOfInterest2 extends Migration
 {
     public function up()
     {
-        Schema::table('bl_regionalinvestment_points_of_interest', function($table)
-        {
-            $table->dropColumn('map_id');
-        });
+        if (Schema::hasColumn('bl_regionalinvestment_points_of_interest', 'map_id')) {
+            Schema::table('bl_regionalinvestment_points_of_interest', function($table)
+            {
+                $table->dropColumn('map_id');
+            });
+        }
     }
     
     public function down()
     {
-        Schema::table('bl_regionalinvestment_points_of_interest', function($table)
-        {
-            $table->bigInteger('map_id')->nullable()->unsigned();
-        });
+        if (!Schema::hasColumn('bl_regionalinvestment_points_of_interest', 'map_id')) {
+            Schema::table('bl_regionalinvestment_points_of_interest', function($table)
+            {
+                $table->bigInteger('map_id')->nullable()->unsigned();
+            });
+        }
     }
 }
