+++ plugins/bl/regionalinvestment/updates/builder_table_update_bl_regionalinvestment_success_stories.php
@@ -7,17 +7,21 @@ class BuilderTableUpdateBlRegionalinvestmentSuccessStories extends Migration
 {
     public function up()
     {
-        Schema::table('bl_regionalinvestment_success_stories', function($table)
-        {
-            $table->bigInteger('map_id')->nullable()->unsigned();
-        });
+        if (!Schema::hasColumn('bl_regionalinvestment_success_stories', 'map_id')) {
+            Schema::table('bl_regionalinvestment_success_stories', function($table)
+            {
+                $table->bigInteger('map_id')->nullable()->unsigned();
+            });
+        }
     }
     
     public function down()
     {
-        Schema::table('bl_regionalinvestment_success_stories', function($table)
-        {
-            $table->dropColumn('map_id');
-        });
+        if (Schema::hasColumn('bl_regionalinvestment_success_stories', 'map_id')) {
+            Schema::table('bl_regionalinvestment_success_stories', function($table)
+            {
+                $table->dropColumn('map_id');
+            });
+        }
     }
 }
