+++ plugins/bl/regionalinvestment/updates/builder_table_update_bl_regionalinvestment_success_stories_2.php
@@ -1,4 +1,6 @@
-<?php namespace BL\RegionalInvestment\Updates;
+<?php
+
+namespace BL\RegionalInvestment\Updates;
 
 use Schema;
 use October\Rain\Database\Updates\Migration;
@@ -7,17 +9,19 @@ class BuilderTableUpdateBlRegionalinvestmentSuccessStories2 extends Migration
 {
     public function up()
     {
-        Schema::table('bl_regionalinvestment_success_stories', function($table)
-        {
-            $table->dropColumn('map_id');
-        });
+        if (Schema::hasColumn('bl_regionalinvestment_success_stories', 'map_id')) {
+            Schema::table('bl_regionalinvestment_success_stories', function ($table) {
+                $table->dropColumn('map_id');
+            });
+        }
     }
-    
+
     public function down()
     {
-        Schema::table('bl_regionalinvestment_success_stories', function($table)
-        {
-            $table->bigInteger('map_id')->nullable()->unsigned();
-        });
+        if (!Schema::hasColumn('bl_regionalinvestment_success_stories', 'map_id')) {
+            Schema::table('bl_regionalinvestment_success_stories', function ($table) {
+                $table->bigInteger('map_id')->nullable()->unsigned();
+            });
+        }
     }
 }
